backend/requirements.txt
amqp==2.6.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
colorama==0.4.3
cx-Oracle==8.0.0
Django==2.2
django-pandas==0.6.2
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.26.1
greenlet==0.4.16
importlib-metadata==1.7.0
isort==5.4.2
jdcal==1.4.1
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
monotonic==1.5
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
pprint==0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.6.0
PyMySQL==0.10.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
vine==1.3.0
wrapt==1.12.1
zipp==3.1.0
